Changes timers to use 32-bit storage, some syntax changes.

+++ PlatformTimer/PlatformTimer.h
@@ -18,7 +18,7 @@
 
 PlatformStatus PlatformTimer_Init( void );
 
-PlatformStatus PlatformTimer_GetTime( uint64_t * const outTime );
+PlatformStatus PlatformTimer_GetTime( uint32_t * const outTime );
 
 PlatformStatus PlatformTimer_Reset( void );
